feat(dashboard): add LevelupChannelConfigComponent
